src/backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
boolean.py==5.0
CacheControl==0.14.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
cyclonedx-python-lib==9.1.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
executing==2.2.0
fastapi==0.121.0
filelock==3.19.1
fonttools==4.60.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.12.0
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
license-expression==30.4.4
lxml==6.0.2
markdown-it-py==4.0.0
markdown2==2.5.4
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
motor==3.7.1
msgpack==1.1.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.5
openai==2.9.0
OTXv2==1.5.12
packageurl-python==0.17.5
packaging==25.0
parso==0.8.4
pillow==11.3.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.3.7
pluggy==1.6.0
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
py-serializable==2.1.0
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.1
pymongo==4.15.1
pyparsing==3.2.3
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pyzmq==26.4.0
requests==2.32.5
rich==14.1.0
ruff==0.14.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
starlette==0.49.1
toml==0.10.2
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.6.0
uvicorn==0.37.0
wcwidth==0.2.13
xlsxwriter==3.2.9
